queryEncode: Do not encode query multiple times.
Multiple encodings on URLs lead to invalid queries to be made
against s3. Handle this by implementing a custom Encode() function.

This is necessary, we have to custom encode characters like '++'
which are erroneously encoded by url.ParseQuery() as space characters
which in turn become '%20' instead the server would be expecting it
to be '%25'

This in turn also fixes - https://github.com/minio/mc/issues/1591