Splitting the main interface into multiple smaller ones
Having a single, all-encompassing interface is ok, but it’s often
useful to pass around a smaller chunk of functionality, to reduce the
surface area of functionality that a function gets access to.

Additionally, passing smaller interfaces around makes it much easier to
write unit tests.

This commit is a first try at splitting up the CloudStorageClient.