retry: Enhance retry to support cases of restricted bucket policy.
This is necessary if GetBucketLocation() is explicitly disabled
for a user policy, fallback to the errorResponse Region and attempt
the request again if necessary. - fixes #332

An example of one such policy

```
{
    "Version": "2012-10-17",
    "Statement": [
        {
            "Effect": "Allow",
            "Action": "s3:ListAllMyBuckets",
            "Resource": "arn:aws:s3:::*"
        },
        {
            "Effect": "Allow",
            "Action": [
                "s3:GetObject",
                "s3:GetObjectAcl",
                "s3:GetObjectVersion",
                "s3:GetObjectVersionAcl",
                "s3:ListBucket"
            ],
            "Resource": [
                "arn:aws:s3:::jarjarbing",
                "arn:aws:s3:::jarjarbing/*"
            ]
        }
    ]
}
```

Clearly user with this policy can list objects on ``jarjarbing`` bucket. But
cannot possibly use our library in this case. This patch addresses this by
using the information returned in the error response and retry the request.