list: Fix listObjects to avoid large bufferred channels to facilitate doneCh.
If we send 1000 bufferred entries simultaneously in bufferred channel, the
top level caller while using doneCh to limit the entries will have no way
to communicate back on how much a ListObjects should be listing the entries
for.

The usual scenario is when one wishes to only get N entires from ListObjects,
It is better for us to send 1 entry at a time so that as and when the
doneCh is called we exit the loop without accumulating more entries than
was needed.

Following program talks about how this issue can happen.
```
func testListBackets() {
    doneCh := make(chan struct{}, 1)
    defer close(doneCh)

    n := 0
    objectCh := s3cli.ListObjects("backet", "", true, doneCh)
    for object := range objectCh {
        if object.Err != nil {
            fmt.Println(object.Err)
            return
        }
        n++
        fmt.Println(n, object)
        if n == 14 {
            doneCh <- struct{}{}
        }
        // waiting here ........................................................
        fmt.Println(n) // Println does not print 14
    }
}
```

Here the program ends up printing 15th object as well even after doneCh indicating
the routine to exit.

Fixes #386