Tweaks to wording and examples
fixed formatting, cleaned up some wording, made the examples more consistent.
Tweaked the first example to not have a compiler error so it will run
Tweaked the file uploader to work more than one time (ignores error on bucket create if the bucket exists)
Added variables for the parameters to make it more clear which is which. Users will probably not be able to tell ID and secret apart. They are a little more complicated but this should be within reason.