api: use simple streaming upload for GetObject() reader.
This change is needed to avoid the offset based ReadAt()
calls which can occur if using io.SectionReader().

Optimizing this part re-implementing a sectionReader()
based on Seeker() wasn't worth the complexity.

This PR also fixes the bug where the streaming signature
was interpreting the Read() error return in a wrong manner.