fix: Mutating variables within go closures should be explicit. (#788)
partSize, a variable on the stack, changes over the course of
uploading multiple objects. This was being used in a function
closure as a go-routine. The variables used in the closure
depend on their values the time of evaluation providing inconsistent
behavior. It is racy to depend on values of variables on stack
at the end of function, when go-routines are executed.

Fixes #784