Support redirect headers to sign again with new Host header.
This is a requirement if say an S3 API endpoint redirects to
a new location based on perhaps a redirect proxy between
two servers.

This is implemented to support bucket re-direction which
might be implemented in future by employing a redirect
proxy.

To test this change please use redirector here -

https://gist.github.com/harshavardhana/a135fafff0d97b31b25ee303ebf28d46