Resign signature and follow redirects
On redirections (503), the default http client will retry the request on
the new Location.  However, the signature will be invalid since it
includes the original Host in the canonical string.

Retry the request after resigning with the new host.  This change
disables the default client retry mechanism and retries the request in
the executeMethod retry loop.  Unfortunately this incurs a retry delay
penalty.