Update functional tests to not expect storage class
Minio server doesn't respond with storage class as object metadata
response, if the storage class is set to STANDARD. This PR updates
the tests to fail if STANDARD is sent as storage class in response
metadata.

Additionally, Minio gateway (except S3 Gateway) currently ignore storage
class requests and hence don't respond with storage class in response
metadata for GetObjectInfo calls. Updated the functional tests to
ignore if storage class is not present in metadata.