Add SSE and refactor encryption API (#942)
* refactor encryption API

This commit does the following changes:

 - Remove all client-side-encryption code
 - Add unified server-side-encryption interface
 - Implement SSE-C and SSE-S3
   (SSE-KMS needs a bit more investigation to finally decide how to implement)
 - Remove special encryption API calls like Get/PutEncryptedObject
   (They are not really usefull after the encryption refactoring)
 - Update functional tests. The functional tests use the new API now.
 - Update API doc and increment version number to 5.0.0
 - Update S3 encryption examples

The client-side-encryption code is removed because it was hardly used,
poorly designed/specified by AWS and has no advantages compared to SSE
for allmost all uses-cases we encounter.

The SSE implementation is located at the `encrypt` package. It exposes
and abstract and easy to use API. Currently the `encrypt` package
implements SSE-C and SSE-S3. SSE-KMS is not yet implemented because some
more investigation is needed to be aware of all details/requirements/implications
of the encryption context which can be passed to SSE-KMS.

All encryption-specific client APIs - like Get/PutEncryptedObject - are
removed because there is no real benefit in keeping them. SSE can simply
be specified as part of Get/Put/Stat ObjectOptions.

The functional tests are updated.

The API documentations is updated.

The S3 encryption examples are updated.
The new examples use the new encryption API.

Fixes #882