tests: Remove partial related tests
testRemovePartiallyUploadedV2(), testListPartiallyUploaded() and
testRemovePartiallyUploaded() were faking an error during uploading
a large file hoping that it would leave some uploaded parts in the server,
but actually minio-go already aborts multipart upload when an error occurs,
so these tests are not useful. Detected when testing function tests using a
NFS backend.