Disable ETag check when copying object < 5 Gb
Currently, copying object is always protected by an ETag check.
This is pretty useful when we copy an object by parts so minio-go
can detect if the object has been changed during the copy operation.
However, this check is not needed if the object size is < 5 Gb.
This change will avoid raising an unnecessary error when an object
has been overwrited by a concurrent S3 client since the resulted
copied object will still be not corrupted.