list: Decode key/prefix and next marker in Core layer (#1169)
Some modules depending on minio-go, such as Minio Gateway S3, depends
on minio-go Core to perform listing. However a regression is found
because decoding keys and common prefixes, along with marker for list v1
is not happening in Core module but in higher level.

This PR moves decoding to Core layer.

This PR also avoids encoding next continuation token since S3 spec
does not require it.