Add Support for Legal-Hold (#1233)
This PR adds support for legal-hold in PutObject, GetObject,
StatObject & CopyObject APIs.

This PR also adds support for "SendContentMd5" field in PutObjectOptions
so users can ask the SDK to send the content-md5. Users are expected
to set this flag for PutObject operations in case locking is enabled
on given bucket.

Note regarding PutObject behaviour:

If SendContentMd5 flag is set minio-go may consume higher memory
due to in-memory md5sum calculation.