fix: close and remove .minio.part file on errors
If a .minio.part file becomes damaged, it lingers in the file
system and can cause future attempts to get the object to fail.
Each failure leaves a file open.

To fix this, if we are returning early from fGetObjectWithContext
(which always returns an error) and we haven't closed the file
safely yet, close and remove it now.