Add custom MD5 hasher option
Since there is no cleanup for the API clients using the SIMD MD5 needs to be up to the API users to manage.

Provide a function that allows to insert a custom hasher.

md5-simd package is only used to provide the interface, by default a wrapper for "crypto/md5" is created, so default performance remains unaffected.