Cache STS Web Identity Credentials
When using the STSWebIdentity credentials provider from the IAM
credentials provider, we need to set the expiry time of the credentials.

We therefore expose the expiry time in the STSWebIdentity credentials
provider so that it can be used by the IAM credentials provider.

Calls to IsExpired() on the IAM credentials provider will then work as
they would have if it had been called on the underlying STSWebIdentity
provider.  Therefore caching the credentials as intended.

Fixes minio/minio-go#1321