Send single DELETE requests if object name has xml invalid chars
Currently, it is not possible to remove some objects using the multi
delete API if those objects contain some weird characters.

In an XML document, some characters are not allowed to be there and
there is no way to escape them in XML 1.0, currently used by S3.
Therefore, some objects that have those invalid characters won't be
deleted in a multi-delete request.

To solve this, we need to check for object names and send individual
DELETE requests. The object name will be encoded in the request URL and
can be removed as expected