sse: re-implement SSE-C, SSE-S3 and SSE-KMS
This commit re-implements the server-side
encryption implementation.

Therefore, the bucket encryption configuration
is moved from `pkg/sse` to `encrypt`.

Further, the `encrypt.ServerSide` interface has
been extended to be more flexible for S3 GET / PUT / COPY
operations.

The reason for this refactoring is that the SSE-KMS implementation
has been broken. It was not possible to specify e.g. the SSE-KMS
context during GET / HEAD operations with the previous abstraction.