Add support for AllVersionsExpiration ilm rule (#2014)
This is a minio-only rule for removing all versions of an
object upon expiration. The rule takes two parameters:
Days and DeleteMarker.
"Days" indicates the number of days after which the object and
all its versions must be expired.
The optional "DeleteMarker" flag indicates that the expiration
be applied if the latest version of this object is a delete marker.