allow users to extend DNS style buckets (#2057)
this is to allow for custom extensions, since
we keep getting new requests to add new URLs
as default for bucket DNS style requests. This
is not scalable for this SDK, allowing a caller
specific function allows callers can write their
own function and provide us when they wish to use
path style or bucket DNS style.