Add support of error responses in 200 OK body (#2115)
An S3 server can respond with 200 OK with an S3 Error response for long
standing calls, such as S3 listing when a prefix contains a lot of
entries.

This commit will support this use case. This can only be enabled with
specific calls that return XML as a successful response, so this will not
be confused with GetObject returning a valid document that contains an
S3 error response.

Supported calls: CompleteMultipartUpload and DeleteMultiObjects.