Allow to set a custom expiry window for preemptive token rotations
Primary use case (from me) is to:
Use the client for to generate presigned urls which are valid for longer.
With the the previous rotate after 80% rule this still cuts away a lot of time of the max. 12h token duration.
This allows the users to use IAM Auth and create presigned urls which can be valid for as long as the expiry window is configured.

Signed-off-by: Jannik Hollenbach <jannik.hollenbach@iteratec.com>