Add post-call checks on AppendObject
Using Append (via PutObject as designed) on a server that doesn't support it can just overwrite the object.. Kinda crappy design by AWS. There is no "preflight" info in HeadObject to indicate if the object is appendable.

Add post-call checks on the expected size to at least indicate to the client there is an issue.